scheduler/pyproject.toml
[project]
name = "scheduler"
version = "0.1.0"
description = ""
authors = [
    {name = "DonAlberton",email = "szpejewski.alberto@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic (>=2.11.2,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "google-cloud-firestore (>=2.20.1,<3.0.0)",
    "google-cloud-pubsub (>=2.29.0,<3.0.0)",
    "google-cloud-tasks (>=2.19.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
